start making it compatible
