+++ primedev/Northstar.cmake
@@ -40,7 +40,9 @@ add_library(NorthstarDLL SHARED
             "core/math/bits.h"
             "core/math/color.cpp"
             "core/math/color.h"
+            "core/math/math_pfns.h"
             "core/math/vector.h"
+            "core/math/vplane.h"
             "core/hooks.cpp"
             "core/hooks.h"
             "core/macros.h"
@@ -111,6 +113,10 @@ add_library(NorthstarDLL SHARED
             "server/auth/serverauthentication.cpp"
             "server/auth/serverauthentication.h"
             "server/alltalk.cpp"
+            "server/ai_helper.cpp"
+            "server/ai_helper.h"
+            "server/ai_navmesh.cpp"
+            "server/ai_navmesh.h"
             "server/buildainfile.cpp"
             "server/r2server.cpp"
             "server/r2server.h"
