+++ primedev/server/ai_helper.h
@@ -1,15 +1,12 @@
 #pragma once
 
-#include "game/server/ai_node.h"
-#include "game/server/ai_navmesh.h"
+#include "server/ai_navmesh.h"
 
 dtNavMesh* GetNavMeshForHull(int nHull);
 
 class CAI_Helper
 {
   public:
-	void SaveNetworkGraph(CAI_Network* pNetwork);
-	void DrawNetwork(CAI_Network* pNetwork);
 	void DrawNavmeshPolys(dtNavMesh* pNavMesh = nullptr);
 };
